pyproject.toml
[tool.poetry]
name = "langraph"
version = "0.1.0"
description = "Learning langgraph and agent building"
authors = ["VishnuRathore98 <rathorevishnu666@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.0"
black = "^25.1.0"
isort = "^6.0.1"
langchain = "^0.3.22"
langgraph = "^0.3.21"
langchain-groq = "^0.3.2"
ipykernel = "^6.29.5"
langchain-openai = "^0.3.12"
langchain-community = "^0.3.21"
langgraph-prebuilt = "^0.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
